+++ bin/proteins-to-pathogens.py
@@ -133,6 +133,12 @@ if __name__ == '__main__':
               'be displayed for that sample.'))
 
     parser.add_argument(
+        '--minProteinCount', type=int, default=0,
+        help=('The minimum number of proteins in a pathogen that must be '
+              'matched by a particular sample in order for that pathogen to '
+              'be displayed for that sample.'))
+
+    parser.add_argument(
         '--pathogenType', default='viral', choices=('bacterial', 'viral'),
         help=('Specify the pathogen type. This option only affects the '
               'language used in HTML output.'))
@@ -253,6 +259,7 @@ if __name__ == '__main__':
         print(grouper.toHTML(
             args.pathogenPanelFilename,
             minProteinFraction=args.minProteinFraction,
+            minProteinCount=args.minProteinCount,
             pathogenType=args.pathogenType,
             title=args.title, preamble=args.preamble,
             sampleIndexFilename=args.sampleIndexFilename,